% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datas.R
\docType{data}
\name{iris}
\alias{iris}
\title{The iris dataset}
\format{
A \code{data.frame} with 150 rows and 5 columns:
\itemize{
\item \code{SepLen}: Sepal Length.
\item \code{SepWid}: Sepal Width.
\item \code{PetLen}: Petal Length.
\item \code{PetWid}: Petal Width.
\item \code{Spe}: Species.
}
}
\source{
This is a copy of the \code{iris} data from the package \code{\link{datasets}}. Run \code{?datasets::iris} to get more details.
}
\usage{
iris
}
\description{
This famous Iris data set!
}
\references{
\url{https://en.wikipedia.org/wiki/Iris_flower_data_set}.
}
\keyword{datasets}
