% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs2.R
\name{normality.df}
\alias{normality.df}
\title{Normality test for a data frame.}
\usage{
normality.df(df, output = c("none", "figure", "message", "all"))
}
\arguments{
\item{df}{data.frame to be tested}

\item{output}{if 'none', nothing is displayed, if 'figure', a figure is displayed, if 'message' a message is displayed in the console and if 'all' a message and a figure are displayed.}
}
\value{
A list with the following elements:
\itemize{
\item \code{pvalues} a vector of the p-values obtained with the Shapiro Wilk test for all numerical variables in the data frame \code{df}.
\item \code{results} a text (printed onto the console, if wanted) describing the results of the performed tests.
}
}
\description{
This function performs the Shapiro Wilk test for all the numerical variables of a given data frame. Each test can be accompanied by a graphical presentation  (density and qqplot) giving a piece of visual information.
}
\examples{
normality.df(iris)

## install.packages("kohonen")
data(wines, package = "kohonen")
normality.df(wines, "all")


}
